\name{EFA.MRFA-internal}
\alias{EFA.MRFA-internal}
\alias{mrfa_s}
\alias{nrm}
\alias{sr}
\alias{adequacymatrix}
\alias{LineCon}
\alias{eigenc}
\alias{eigend}
\alias{fitchi}
\alias{fitchi_zero}
\alias{ml77}
\alias{tri}
\alias{size}
\alias{transpose}
\title{
EFA.MRFA-internal
}
\description{
This functions are for internal use only.
}
\usage{
mrfa_s
nrm
sr
adequacymatrix
LineCon
eigenc
eigend
fitchi
fitchi_zero
ml77
tri
size
transpose
}
\author{
David Navarro-Gonzalez

Urbano Lorenzo-Seva
}
\keyword{ internal }
